DEMO01/requirements.txt
opencv-python
mediapipe
numpy
scikit-learn
joblib
gTTS
python-vlc